The wave system is finally here

+++ src/afyber/shmupfeaturecreep/engine/world/Global.java
@@ -11,14 +11,14 @@ public class Global {
 	private Global() {}
 
 	private static final HashMap<String, Integer> allIntGlobals = new HashMap<>();
-	private static final HashMap<String, Float> allFloatGlobals = new HashMap<>();
+	private static final HashMap<String, Double> allFloatGlobals = new HashMap<>();
 	private static final HashMap<String, String> allStringGlobals = new HashMap<>();
 
 	public static void setIntGlobal(String name, int value) {
 		allIntGlobals.put(name, value);
 	}
 
-	public static void setFloatGlobal(String name, float value) {
+	public static void setFloatGlobal(String name, double value) {
 		allFloatGlobals.put(name, value);
 	}
 
@@ -30,7 +30,7 @@ public class Global {
 		return allIntGlobals.get(name);
 	}
 
-	public static float getFloatGlobal(String name) {
+	public static double getFloatGlobal(String name) {
 		return allFloatGlobals.get(name);
 	}
